requirements.txt
 argon2-cffi==18.3.0
 celery==4.2.1
 coverage==4.5.2
 beautifulsoup4==4.6.3
 bokeh==1.0.2
 Django==2.1.3
 django-extensions==2.1.3
 djangorestframework==3.9.0
 django-celery-results==1.0.4
 django-celery-beat==1.3.0
 django-guardian==1.4.9
 django-picklefield==2.0
 django-select2==6.3.1
 django-tables2==2.0.3
 django-nose==1.4.6
 django-widget-tweaks==1.4.3
 factory_boy==2.11.1
 ipdb==0.10.3
 matplotlib==3.0.2
 psutil==5.4.8
 psycopg2==2.7.6.1
 pyomo==5.5.1
 python-decouple==3.1
 quandl==3.4.5
 redis==3.0.1
 selenium==3.141.0
 scikit-learn==0.20.0
 splinter==0.10.0
 statsmodels==0.9.0
 watchdog==0.9.0
